requirements.txt
# Core dependencies
pydantic>=2.5.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Validation & diffing
deepdiff>=6.7.0
jsonschema>=4.20.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
mypy>=1.7.0
